Rename demo-app from FCM App to Remote Notify Demo App with proper git history
- Use git mv to preserve file history when moving Kotlin files
- Change package name from org.nella.fcmapp to org.nella.RN.demo (capital RN)
- Update applicationId and namespace in build.gradle to new package name
- Rename FCMService to RnDemoService class and update all references
- Update AndroidManifest.xml with new service name and theme references
- Change app name from 'FCM App' to 'Remote Notify Demo App' in strings.xml
- Update theme name from Theme.FCMApp to Theme.RemoteNotifyDemoApp
- Replace user-facing 'FCM' references with 'notification' terminology
- Change default backend URL to https://demo.rn.nella.org
- Update README.md with new app name, package references, and default URL
- Maintain Firebase-specific technical terms where appropriate
- Update SharedPreferences name from FCMAppSettings to RnDemoSettings
- Change notification channel ID from fcm_default_channel to rn_demo_channel
- Update all UI hints and examples to use new demo.rn.nella.org domain
- Preserve all Firebase Cloud Messaging functionality and dependencies
