+++ demo-app/README.md
@@ -1,6 +1,6 @@
-# FCM Android Demo App
+# Remote Notify Demo App
 
-Android client that registers FCM tokens using hybrid encryption and receives push notifications.
+Android client that registers notification tokens using hybrid encryption and receives push notifications.
 
 ## Setup
 
@@ -8,7 +8,7 @@ Android client that registers FCM tokens using hybrid encryption and receives pu
 
 1. Go to [Firebase Console](https://console.firebase.google.com/)
 2. Create or select a project
-3. Add Android app with package name: `org.nella.fcmapp` (demo app)
+3. Add Android app with package name: `org.nella.RN.demo` (demo app)
 4. Download `google-services.json` and replace the placeholder in `app/` directory
 
 ### 2. Public Key
@@ -20,11 +20,11 @@ Ensure `app/src/main/assets/public_key.pem` contains the RSA public key (see mai
 1. Open project in Android Studio
 2. Sync project to download dependencies
 3. Run on device or emulator
-4. Tap "Register Device Token" to encrypt and send FCM token
+4. Tap "Register Device Token" to encrypt and send notification token
 
 ## Features
 
-- Single-button FCM token registration
+- Single-button notification token registration
 - Client-side hybrid encryption (AES-256-GCM + RSA-4096)
 - Automatic certificate bypass for development servers
 - Push notification handling
@@ -45,7 +45,7 @@ Sends encrypted token to `https://10.0.2.2:8443/register` (emulator) with payloa
 
 ## Dependencies
 
-- Firebase Cloud Messaging: `com.google.firebase:firebase-messaging-ktx:23.3.1`
+- Firebase Cloud Messaging: `com.google.firebase:firebase-messaging-ktx:24.1.2`
 - OkHttp: `com.squareup.okhttp3:okhttp:4.12.0`
 - AndroidX UI components
