+++ demo-app/app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.FCMApp"
+        android:theme="@style/Theme.RemoteNotifyDemoApp"
         android:networkSecurityConfig="@xml/network_security_config"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.FCMApp">
+            android:theme="@style/Theme.RemoteNotifyDemoApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -30,11 +30,11 @@
             android:name=".SettingsActivity"
             android:exported="false"
             android:parentActivityName=".MainActivity"
-            android:theme="@style/Theme.FCMApp" />
+            android:theme="@style/Theme.RemoteNotifyDemoApp" />
         
         <!-- Firebase Messaging Service -->
         <service
-            android:name=".FCMService"
+            android:name=".RnDemoService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
