+++ demo-app/app/src/main/java/org/nella/RN/demo/RnDemoService.kt
@@ -1,4 +1,4 @@
-package org.nella.fcmapp
+package org.nella.RN.demo
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -11,11 +11,11 @@ import androidx.core.app.NotificationCompat
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 
-class FCMService : FirebaseMessagingService() {
+class RnDemoService : FirebaseMessagingService() {
     
     companion object {
-        private const val TAG = "FCMService"
-        private const val CHANNEL_ID = "fcm_default_channel"
+        private const val TAG = "RnDemoService"
+        private const val CHANNEL_ID = "rn_demo_channel"
     }
     
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
@@ -26,7 +26,7 @@ class FCMService : FirebaseMessagingService() {
         // Check if message contains a notification payload
         remoteMessage.notification?.let {
             Log.d(TAG, "Message Notification Body: ${it.body}")
-            sendNotification(it.title ?: "FCM Message", it.body ?: "")
+            sendNotification(it.title ?: "Notification", it.body ?: "")
         }
         
         // Check if message contains a data payload
@@ -40,7 +40,7 @@ class FCMService : FirebaseMessagingService() {
         
         // If you want to send messages to this application instance or
         // manage this app's subscriptions on the server side, send the
-        // FCM registration token to your app server.
+        // Firebase registration token to your app server.
         sendRegistrationToServer(token)
     }
     
@@ -68,7 +68,7 @@ class FCMService : FirebaseMessagingService() {
         // Since android Oreo notification channel is needed.
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             val channel = NotificationChannel(channelId,
-                "FCM Channel",
+                "RN Demo Channel",
                 NotificationManager.IMPORTANCE_DEFAULT)
             notificationManager.createNotificationChannel(channel)
         }
