+++ demo-app/app/src/main/res/layout/activity_settings.xml
@@ -24,7 +24,7 @@
         android:id="@+id/urlEditText"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:hint="https://r-notify.nella.org"
+        android:hint="https://demo.rn.nella.org"
         android:inputType="textUri"
         android:layout_marginBottom="16dp"
         android:padding="12dp"
@@ -33,7 +33,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Examples:\n• https://r-notify.nella.org\n• https://your-domain.com\n• https://10.0.2.2:8443 (Android emulator)"
+        android:text="Examples:\n• https://demo.rn.nella.org\n• https://your-domain.com\n• https://10.0.2.2:8443 (Android emulator)"
         android:textSize="14sp"
         android:textColor="@android:color/darker_gray"
         android:layout_marginBottom="24dp" />
@@ -66,7 +66,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="\n🔒 Security Note:\nAll communication uses encrypted FCM tokens. The backend URL should use HTTPS for security."
+        android:text="\n🔒 Security Note:\nAll communication uses encrypted notification tokens. The backend URL should use HTTPS for security."
         android:textSize="12sp"
         android:textColor="@android:color/darker_gray"
         android:layout_marginTop="24dp"
