+++ demo-app/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">FCM App</string>
+    <string name="app_name">Remote Notify Demo App</string>
 </resources>
